requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
opencv-python==3.4.0.12
Pillow==5.1.0
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.7.2
pytz==2018.4
scipy==1.0.1
six==1.11.0
tesseract==0.1.3
tesserocr==2.2.2
